// src/types/api/objective.ts
/**
 * LifeMapWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {ObjectiveStakeholder} from "~/types/api/objectiveStakeholder";
import {ActionItem} from "~/types/api/actionItem";
import {Person} from "./person";
import {MetricOfSuccess} from "~/types/api/metricOfSuccess";
import {DimensionOfLife} from "~/types/api/dimensionOfLife";


export interface Objective { 
    ObjectiveID: number;
    HouseholdID?: number;
    PersonID?: number;
    ClientID?: number;
    InterviewID: number;
    QuestionID?: number;
    InterviewResponseID?: number;
    DimensionOfLifeID?: number;
    MetricOfSuccessID?: number;
    Description: string;
    Champion?: number;
    OtherPeople?: string;
    MLPA?: boolean;
    PriorityNow?: boolean;
    PriorityLater?: boolean;
    DIY?: boolean;
    AssistanceNeeded?: boolean;
    Rank?: number;
    Importance?: number;
    IsHidden?:boolean;
    SuccessImage?: number;
    SuccessDescription?: string;
    SuccessImageURL?: string;
    StartDate?: Date;
    StartDateType?: number;
    ProjectedEndDate?: Date;
    EndDateType?: number;
    ScheduleFrequencyID?: number;
    ScheduleDetail?: string;
    AmountOfTimeNeeded?: string;
    KnowledgeYesNo?: number;
    KnowledgeNow?: number;
    KnowledgeNeeded?: number;
    KnowledgeNeedToKnow?: string;
    KnowledgeTask?: number;
    KnowledgeAdvisorHelp?: string;
    PersonalImpactLevel?: number;
    PersonalExperience?: string;
    ExperienceNow?: number;
    ExperienceNeeded?: number;
    FundingKnown?: number;
    TotalFundingAmount?: number;
    InstallmentFrequency?: number;
    InstallmentFrequencyOther?: string;
    InstallmentAmount?: number;
    FundingDetail?: string;
    PersonalBalanceSheet?: number;
    FamilyBalanceSheet?: number;
    Connections?: string;
    StatusID?: number;
    StatusText?: string;
    TimeframeID?: number;
    CreationDate?: string;
    CreatedBy?: number;
    LastModifiedDate?: string;
    LastModifiedBy?: number;
    PercentComplete?: number;

    Person?: Person;
    ChampionPerson?: Person;
    Stakeholders?: ObjectiveStakeholder[];
    ActionItemList?: ActionItem[];
    MetricOfSuccess?: MetricOfSuccess;
    DimensionOfLife?: DimensionOfLife;
    Why?: string;
}